# uploads/openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "PhutiCee APIs",
      "description": "This is Swagger Documetation",
      "version": "1.0.0",
      "contact": {}
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "paths": {
      "/api/v1/auth/register": {
        "post": {
          "summary": "Register User API",
          "description": "",
          "operationId": "postapiv1authregister",
          "parameters": [],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "example": {}
                  }
                }
              }
            }
          },
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "name": "User 8",
                    "email": "user8@gmail.com",
                    "password": "12345678",
                    "gender": "male"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/auth/login": {
        "post": {
          "summary": "Login User API",
          "description": "",
          "operationId": "postapiv1authlogin",
          "parameters": [],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "example": {}
                  }
                }
              }
            }
          },
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "email": "user8@gmail.com",
                    "password": "12345678"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/auth/profile": {
        "get": {
          "summary": "User Profile API",
          "description": "",
          "operationId": "getapiv1authprofile",
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "schema": {
                "type": "string"
              },
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY4MGRmMGRhM2M1Mjc2MDgzZDY2ZmJmOCIsIm5hbWUiOiJVc2VyIDgiLCJlbWFpbCI6InVzZXI4QGdtYWlsLmNvbSIsImlhdCI6MTc0NTc0NDE1NiwiZXhwIjoxNzQ2MzQ4OTU2fQ.BUb1-AjnsWHNHppBzEHXuuKrRaOxaX5Q-Ev0sWCBTFc"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "example": {}
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
...
      "securitySchemes": {
        "basicAuth": {
          "type": "http",
          "scheme": "basic"
        },
        "digestAuth": {
          "type": "http",
          "scheme": "digest"
        }
      }
    },
...
  }